+++ example/lib/custom_icon_pack.dart
@@ -1,48 +1,38 @@
 import 'package:flutter/material.dart';
-import 'package:flutter_iconpicker/Models/icon_pack.dart';
 import 'package:flutter_iconpicker/Models/icon_picker_icon.dart';
 
 const Map<String, IconPickerIcon> myCustomIcons = {
   'threesixty': IconPickerIcon(
-      name: 'threesixty', data: Icons.threesixty, pack: IconPack.custom),
+      name: 'threesixty', data: Icons.threesixty, pack: 'custom'),
   'threed_rotation': IconPickerIcon(
-      name: 'threed_rotation',
-      data: Icons.threed_rotation,
-      pack: IconPack.custom),
-  'four_k':
-      IconPickerIcon(name: 'four_k', data: Icons.four_k, pack: IconPack.custom),
-  'ac_unit': IconPickerIcon(
-      name: 'ac_unit', data: Icons.ac_unit, pack: IconPack.custom),
+      name: 'threed_rotation', data: Icons.threed_rotation, pack: 'custom'),
+  'four_k': IconPickerIcon(name: 'four_k', data: Icons.four_k, pack: 'custom'),
+  'ac_unit':
+      IconPickerIcon(name: 'ac_unit', data: Icons.ac_unit, pack: 'custom'),
   'access_alarm': IconPickerIcon(
-      name: 'access_alarm', data: Icons.access_alarm, pack: IconPack.custom),
+      name: 'access_alarm', data: Icons.access_alarm, pack: 'custom'),
   'access_alarms': IconPickerIcon(
-      name: 'access_alarms', data: Icons.access_alarms, pack: IconPack.custom),
+      name: 'access_alarms', data: Icons.access_alarms, pack: 'custom'),
   'access_time': IconPickerIcon(
-      name: 'access_time', data: Icons.access_time, pack: IconPack.custom),
+      name: 'access_time', data: Icons.access_time, pack: 'custom'),
   'accessibility': IconPickerIcon(
-      name: 'accessibility', data: Icons.accessibility, pack: IconPack.custom),
+      name: 'accessibility', data: Icons.accessibility, pack: 'custom'),
   'accessibility_new': IconPickerIcon(
-      name: 'accessibility_new',
-      data: Icons.accessibility_new,
-      pack: IconPack.custom),
+      name: 'accessibility_new', data: Icons.accessibility_new, pack: 'custom'),
   'accessible': IconPickerIcon(
-      name: 'accessible', data: Icons.accessible, pack: IconPack.custom),
+      name: 'accessible', data: Icons.accessible, pack: 'custom'),
   'accessible_forward': IconPickerIcon(
       name: 'accessible_forward',
       data: Icons.accessible_forward,
-      pack: IconPack.custom),
+      pack: 'custom'),
   'account_balance': IconPickerIcon(
-      name: 'account_balance',
-      data: Icons.account_balance,
-      pack: IconPack.custom),
+      name: 'account_balance', data: Icons.account_balance, pack: 'custom'),
   'account_balance_wallet': IconPickerIcon(
       name: 'account_balance_wallet',
       data: Icons.account_balance_wallet,
-      pack: IconPack.custom),
+      pack: 'custom'),
   'account_box': IconPickerIcon(
-      name: 'account_box', data: Icons.account_box, pack: IconPack.custom),
+      name: 'account_box', data: Icons.account_box, pack: 'custom'),
   'account_circle': IconPickerIcon(
-      name: 'account_circle',
-      data: Icons.account_circle,
-      pack: IconPack.custom),
+      name: 'account_circle', data: Icons.account_circle, pack: 'custom'),
 };