+++ lib/IconPicker/search_bar.dart
@@ -4,12 +4,11 @@
 /// rebar.ahmad@gmail.com
 
 import 'package:flutter/material.dart';
+import 'package:flutter_iconpicker/Helpers/icon_pack_manager.dart';
 import 'package:flutter_iconpicker/Models/icon_picker_icon.dart';
 import 'package:flutter_iconpicker/controllers/icon_controller.dart';
 import 'package:provider/provider.dart';
 import '../Models/icon_pack.dart';
-import '../Helpers/color_brightness.dart';
-import 'icons.dart';
 
 class FIPSearchBar extends StatefulWidget {
   const FIPSearchBar({
@@ -47,12 +46,13 @@ class _FIPSearchBarState extends State<FIPSearchBar> {
   void _search(String searchValue) {
     List<MapEntry<String, IconPickerIcon>> searchResult = [];
 
-    final flatFIPPacks = widget.iconPack
-            ?.expand((pack) => FIPIconManager.getSelectedPack(pack).entries) ??
+    final Iterable<MapEntry<String, IconPickerIcon>> flatFIPPacks = widget
+            .iconPack
+            ?.expand((pack) => IconPackManager.getIcons(pack).entries) ??
         [];
     final flatCustomPacks = widget.customIconPack?.entries ?? [];
 
-    for (var item in flatFIPPacks) {
+    for (MapEntry<String, IconPickerIcon> item in flatFIPPacks) {
       if (searchComparator.call(
           searchValue,
           IconPickerIcon(
@@ -65,7 +65,9 @@ class _FIPSearchBarState extends State<FIPSearchBar> {
       if (searchComparator.call(
           searchValue,
           IconPickerIcon(
-              name: item.key, data: item.value.data, pack: IconPack.custom))) {
+              name: item.key,
+              data: item.value.data,
+              pack: IconPack.custom.name))) {
         searchResult.add(item);
       }
     }
@@ -85,18 +87,8 @@ class _FIPSearchBarState extends State<FIPSearchBar> {
       return TextField(
         onChanged: (val) => _search(val),
         controller: controller.searchTextController,
-        style: TextStyle(
-          color: FIPColorBrightness(widget.backgroundColor!).isLight()
-              ? Colors.black
-              : Colors.white,
-        ),
         decoration: InputDecoration(
           contentPadding: const EdgeInsets.only(top: 15),
-          hintStyle: TextStyle(
-            color: FIPColorBrightness(widget.backgroundColor!).isLight()
-                ? Colors.black54
-                : Colors.white54,
-          ),
           hintText: widget.searchHintText,
           prefixIcon: widget.searchIcon,
           suffixIcon: AnimatedSwitcher(
@@ -112,8 +104,7 @@ class _FIPSearchBarState extends State<FIPSearchBar> {
 
                       if (widget.iconPack != null) {
                         for (var pack in widget.iconPack!) {
-                          controller
-                              .addAll(FIPIconManager.getSelectedPack(pack));
+                          controller.addAll(IconPackManager.getIcons(pack));
                         }
                       }
                     }),