+++ lib/extensions/map_extensions.dart
@@ -1,0 +1,4 @@
+extension NullableMapExtensions on Map? {
+  bool get isNullOrEmpty => this == null || this!.isEmpty;
+  bool get isNotNullOrEmpty => !isNullOrEmpty;
+}